+++ spec/whois/server/adapters/formatted_spec.rb
@@ -5,25 +5,26 @@ describe Whois::Server::Adapters::Formatted do
   let(:definition) { [:tld, ".de", "whois.denic.de", { :format => "-T dn,ace -C US-ASCII %s" }] }
 
 
-  describe "#query" do
+  describe "#lookup" do
     it "returns the WHOIS record" do
       response = "Whois Response"
       expected = response
       server = klass.new(*definition)
       server.query_handler.expects(:call).with("-T dn,ace -C US-ASCII domain.de", "whois.denic.de", 43).returns(response)
 
-      record = server.query("domain.de")
+      record = server.lookup("domain.de")
       record.to_s.should  == expected
       record.parts.should == [Whois::Record::Part.new(:body => response, :host => "whois.denic.de")]
     end
 
     context "without format option" do
       it "raises an error" do
-        lambda do
-          server = klass.new(*[:tld, ".de", "whois.denic.de", {}])
-          server.query_handler.expects(:call).never
-          server.query("domain.de")
-        end.should raise_error(Whois::ServerError)
+        server = klass.new(*[:tld, ".de", "whois.denic.de", {}])
+        server.query_handler.expects(:call).never
+
+        expect {
+          server.lookup("domain.de")
+        }.to raise_error(Whois::ServerError)
       end
     end
 
@@ -33,7 +34,7 @@ describe Whois::Server::Adapters::Formatted do
         server = klass.new(:tld, ".de", "whois.denic.de", { :format => "-T dn,ace -C US-ASCII %s", :port => 20 })
         server.query_handler.expects(:call).with("-T dn,ace -C US-ASCII domain.de", "whois.denic.de", 20).returns(response)
 
-        server.query("domain.de")
+        server.lookup("domain.de")
       end
     end
 
@@ -44,7 +45,7 @@ describe Whois::Server::Adapters::Formatted do
         server.configure(:bind_host => "192.168.1.1", :bind_port => 3000)
         server.query_handler.expects(:call).with("-T dn,ace -C US-ASCII domain.de", "whois.denic.de", 43, "192.168.1.1", 3000).returns(response)
 
-        server.query("domain.de")
+        server.lookup("domain.de")
       end
     end
   end